# public/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Always Scheming Terminal API",
    "version": "1.0.0",
    "description": "OpenAPI specification for AST API v1 endpoints. Last verified: 2026-07-24"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Relative to base URL"
    }
  ],
  "paths": {
...
    "/entity-aliases": {
      "get": {
        "summary": "Get alias map or search entities",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Search prefix for entity canonical names (autocomplete)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alias map or search results",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "results"
                      ]
                    },
                    {
                      "type": "object",
                      "properties": {
                        "aliases": {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        }
                      },
                      "required": [
                        "aliases"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new entity alias (admin)",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "alias": {
                    "type": "string"
                  },
                  "canonical_name": {
                    "type": "string"
                  }
                },
                "required": [
                  "alias",
                  "canonical_name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alias created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "alias": {
                      "type": "string"
                    },
                    "entity": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "alias",
                    "entity"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Alias already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key"
      }
    }
  },
  "security": []
}